Bug fix : - Correction d'échelle dans la boite slide - Réglage des arrondis dans les calculs d'image - Ajout de l'espagnol au fichier .desktop (icone linux)
git-svn-id: file:///media/werkstatt/064f12d4-5f27-4839-9537-9ffb7939d236/ffDiaporama-Entwicklung-2023/ffdiaporama-svn/trunk@124 f803d417-eb43-4c80-8848-1a7435ce69a7

+++ ffDiaporama/_ImagesDefinitions.cpp
@@ -203,7 +203,7 @@ QImage *cFilterCorrectObject::GetImage(QImage *LastLoadedImage,int Width,int Hei
     PB.fillRect(QRect(0,0,Width+1,Height+1),Qt::transparent);
     PB.setCompositionMode(QPainter::CompositionMode_SourceOver);
 
-    PB.drawImage(QRectF(DstX,DstY,DstW,DstH),*SourceImage,QRectF((RealImageW/2)-(Hyp/2)+SrcX,(RealImageH/2)-(Hyp/2)+SrcY,SrcW,SrcH));
+    PB.drawImage(QRectF(DstX,DstY,DstW,DstH),*SourceImage,QRectF(int((RealImageW/2)-(Hyp/2)+SrcX),int((RealImageH/2)-(Hyp/2)+SrcY),SrcW,SrcH));
 
     PB.end();
